Deadlock_feather/LLM/requirements.txt
llama-cpp-python
psutil
rich